Add DC additions and subtractions variables
