+++ policyengine_us/parameters/gov/states/dc/tax/income/additions/sources.yaml
@@ -1,7 +1,7 @@
-description: DC subtracts these items US AGI when computing DC AGI.
+description: DC adds these items to US AGI when computing DC AGI.
 metadata:
   unit: variable
-  label: DC subtractions from US adjusted gross income
+  label: DC additions to US adjusted gross income
   reference:
     - title: 2021 Form D40 Booklet, Schedule I
       href: https://otr.cfo.dc.gov/sites/default/files/dc/sites/otr/publication/attachments/52926_D-40_12.21.21_Final_Rev011122.pdf#page=63
@@ -9,5 +9,4 @@ metadata:
       href: https://otr.cfo.dc.gov/sites/default/files/dc/sites/otr/publication/attachments/2022_D-40_Booklet_Final_blk_01_23_23_Ordc.pdf#page=55
 values:
   2021-01-01:
-    - dc_disability_income_subtraction
-    - dc_disabled_exclusion_subtraction
+    - dc_self_employment_loss_addition
